Align KubeExec with CommandClass style format (#1857)
* Separate message log validation

Client code is responsible for calling this to validate messages before
saving them to a MessageLog instance. This separation allows validation
to be handled as needed, while keeping MessageLog a pure value object.

* Rename file before refactor

This will keep the git history clean as we won't have a deleted
and created file.

* Align KubeExec with CommandClass style format

- Rename class to the action it performs
- Use dependencies so we can easily mock them
- call function tells the class's story

This commit also moves the error validation of websocket errors
into the `KubeExec` (renamed to `ExecuteCommandInContainer`) class.

Prior to this commit, the KubeExec class didn't verify that
there were no errors in the stream and returned the messages to
the caller. Then, the caller (in this case 'InjectClientCert')
parsed the messages and verified there were no errors. Instead,
it is better to handle the errors in the KubeExec class and raise
an error if one occurred.

* Extract WebSocket Client events into separate class

The event handling has a lot of logic which makes the
ExecuteCommandInContainer class pretty big. Moving this
logic to a separate class will make both classes more
readable.

* Improve logging around cert validation

- Add stack trace to test failures
- Continue polling for cert file in case of Docker error

* Add UTs for ExecuteCommandInContainer

* Verify that cert injection errors are written to the client

This commit adds a deployment where the authenticator container
is an image that is the same as the others only that it doesn't
have the directory /etc/conjur/ssl. This will cause the cert
injection to fail so we can verify that the injection error
is written to the container

* Add log tags to Kubectl exec Websocket logs

Connected to #1840

The WebSocket client event handlers run in a sub-thread and the log
tags (origin, thread id, etc.) are not passed automatically. This commit
adds the log tags from the main thread so the WebSocket logs have them.

We also append the sub-thread id to the main one so we can easily know the
flow from the logs and connect between the threads.

So instead of having this log:
```
[origin=10.1.3.124] [request_id=07e98cd9-8181-4a2e-a57e-ca30b587b62c] [tid=35] CONJ00015D Copying SSL certificate to cyberark-secrets-provider-for-k8s:/etc/conjur/ssl/client.pem in local-secrets-provider/test-env-6b7b4b57c6-f8xbg
CONJ00010D Pod 'test-env-6b7b4b57c6-f8xbg' : channel open
CONJ00012D Pod 'test-env-6b7b4b57c6-f8xbg', channel 'stdout':
CONJ00013D Pod: 'test-env-6b7b4b57c6-f8xbg', message: 'close', data: 'nil'
CONJ00011D Pod 'test-env-6b7b4b57c6-f8xbg' : channel closed
[origin=10.1.3.124] [request_id=07e98cd9-8181-4a2e-a57e-ca30b587b62c] [tid=35] CONJ00037D Copied SSL certificate successfully
```

we now have this log:
```
[origin=10.1.4.168] [request_id=341a0cb2-242d-4b46-9f7d-c2f79a2c8d41] [tid=33] CONJ00015D Copying client certificate to cyberark-secrets-provider-for-k8s:/etc/conjur/ssl/client.pem in local-secrets-provider/test-env-6b7b4b57c6-mkpxk
[origin=10.1.4.168] [request_id=341a0cb2-242d-4b46-9f7d-c2f79a2c8d41] [tid=33=>77] CONJ00010D Pod 'test-env-6b7b4b57c6-mkpxk' : channel open
[origin=10.1.4.168] [request_id=341a0cb2-242d-4b46-9f7d-c2f79a2c8d41] [tid=33=>77] CONJ00012D Pod 'test-env-6b7b4b57c6-mkpxk', channel 'stdout':
[origin=10.1.4.168] [request_id=341a0cb2-242d-4b46-9f7d-c2f79a2c8d41] [tid=33=>77] CONJ00013D Pod 'test-env-6b7b4b57c6-mkpxk': message: 'close', data: 'nil'
[origin=10.1.4.168] [request_id=341a0cb2-242d-4b46-9f7d-c2f79a2c8d41] [tid=33=>77] CONJ00011D Pod 'test-env-6b7b4b57c6-mkpxk' : channel closed
[origin=10.1.4.168] [request_id=341a0cb2-242d-4b46-9f7d-c2f79a2c8d41] [tid=33] CONJ00037D Started copying the client certificate successfully
```

Some other changes:
- Replace sleep polling with queue
- Specify units in default timeout
- Move timeout warn to logs.rb
- Raise error instead of adding it to stream in case of handshake

Co-authored-by: Jonah <jonahx@gmail.com>