Make env args robust against spaces
This commit does a few bits of cleanup, but the primary purpose is to
remove a hidden assumption, a bug waiting to happen, by making the
env arguments we pass to cucumber robust against spaces.

Previously, our code was working only by luck: If one of the env
argument values that we were passing to cucumber had happened to contain
a space, the code would break.  It happened to be the case that none of
them did, but this might change at any time.

In order to fix this, this commit changes the functions responsible for
constructing the env args from ones that return strings to ones that
"return" arrays.  I put return in scare quotes because you cannot return
an array in bash, but you can use a bash nameref to populate a given
argument as an array.

To signify this change, the function names have changed from "_get_XXX"
to "_hydrate_XXX", and the functions which use them have been updated
appropriately.

Additional cleanup:

- Simplify how "_hydrate_gcp_env_args" works:  Replace multiple calls to
  "jq" with a single call, and eliminate a bunch of unnecessary and
  cluttery temp variables by using a array.
- Remove superfluous echo statements.
- Add comments.
- Improve formatting.