Replace : DOC strings with standard help function
Currently the CLI documentation in the "help" function is generated by
meta-code that reads the source code of "ci/test" itself and parses it
to find lines like:

  : DOC Some help message

This is extremely brittle and non-standard and was already the source of
two bugs I found.

This commit removes these DOC strings and their associated helper
functions, and replaces them with a standard hard-coded bash help
function.