Rotating the API key of an nonexistent role now returns an error code
This fixes a bug where a user rotating a nonexistent role's API key with basic auth would
receive a successful status code implying that the nonexistent role's key was rotated. In
reality, the user was rotating their own API key. The action of rotating a nonexistent user's
API key is now rejected with a 401 Unauthorized status code. Adds test cases to confirm.