Refactor ci/test to decouple individual tests
This commit is a complete refactor of ci/test whose main purpose is to
decouple the "test runner" code from the individual tests.  This makes
the bash much easier to understand and maintain, and ensures that every
individual test is independent, and that its dependencies are clear.

This was accomplished by:

1. Stripping down "ci/test" to its bare essentials.  It now only
   contains 4 functions: main, help, finish (for cleanup work), and
   docker_diagnositics (for printing docker logs).

2. Creating a folder "ci/test_suites", whose subfolders are the
   individual tests.  The entrypoint for every test is simply the
   file named "test" in its folder.

3. Putting common functions shared by the tests into a "shared.sh" file,
   which can be sourced by the individual as needed.

4. Updating the "Jenkinsfile" to reference the tests using this new
   convention.