Add missing shift in CI image push script
In the PR to add a UBI based image (https://github.com/cyberark/conjur/pull/1927),
we added a new argument to the `tag_and_push` function in our Docker image
push step for CI. However, this addition neglected to shift the remaining
arguments. This lead the function to interpret the source image in the list of
image targets to push to, leading to a malformed image reference.