Fix issues with running push-image.sh
This commit updates the scripts that reference version_utils.sh to use
shebang `/usr/bin/env bash` instead of `/bin/bash`, so that users with different
locations for bash can all use these same scripts.

It also updates push-image.sh so that the Summon invocation happens in the
Jenkinsfile instead of in the script directly, and fixes the bug in the
`readarray` bash logic that would cause this script to hang in some cases.