Don't fail if CodeClimate cannot be prepared
Prevents a build failure if preparations for submitting a report to
CodeClimate cannot be completed, primarily due to builds occuring
outside of the main repository.  Results from the build submitted to
CodeClimate can still fail the build, but a failure of the preparation
stage will prevent the attempt to submit a report.