Always run nightly tests
Recently, we introduced a change to skip running the CI tests when the
only changed files were markdown files, so that design doc and other
non-code PRs wouldn't block on the CI.

However, this had the unintended side-effect of skipping our nightly
builds, because the criteria for "only markdown files":

    git diff origin/master --name-only |
    grep -v "^.*\\.md$" > /dev/null'

will also _always_ be true when a nightly build is auto-triggered.

This commit ensures that the nightly builds will always be run.