Update image publish script to strip "v" from tag
In recent updates to the publish script, we started using the Jenkins TAG_NAME instead of the VERSION from the repo. This had the unfortunate side effect of publishing images with Docker tags of the form "v1.*", instead of publishing images without the "v" prefix. This commit updates the logic in the file so that it will push images without the "v" prefix, as expected.