Verify ID token as part of the OIDC authenticator specific validations
We actually don't have any specific validations for OIDC. We only verify
that the ID token is valid but this is done while it is decoded (using
a third-party). However, we want to verify that we verify the token no
matter what so we run the validation again (even if it means that in most
cases we will perform this action twice).