Avoid using app errors in ConjurConfig
ConjurConfig is consumed by conjurctl, which is run without Rails auto-loading.
This means that in order to use the errors defined in the application code in
ConjurConfig, we have to add several require statements to the top of any
conjurctl commands that rely on ConjurConfig OR we have to add requires to the
application classes, which is discouraged in Rails and causes a flood of
warnings to occur whenever Rails starts.

To dodge this issue entirely (for now), we are going to temporarily isolate
ConjurConfig from any application code by using an error class defined in
ConjurConfig itself. We will hook into the error system again when we have a
good way to share code between the Rails application and conjurctl.