Remove conjur-ruby-api dependency from tests
Previously, we used conjur-ruby-api (an external gem) in our cucumber
tests to test drive the API.  This caused a problem when we changed the
API -- eg, by adding a new endpoint: In order to test your change, you
had to _first_ update the external gem so that it knows about your new
endpoint.  But that gem shouldn't be updated until the new Conjur
endpoint is tested and released.  This made testing very difficult.

This commit solves the problem by removing the dependency on
conjur-ruby-api.  Instead, we now test our own API simply by using
rest-client.