Move HTTP response normalization into Client
Prior to this commit, we used a cucumber helper `api_response` to
normalize http responses made by `Client`.  But this responsibility
belongs in Client itself.  Here we introduce a thin wrapper over
`RestClient` (`RestClientWrapper`) that delegates to RestClient but
rescues errors and converts them into objects with the same interface as
successful responses.

This lets us update Client with a just a tiny change, and since we
update the `uri` method, it ensures that any future methods added to
Client will automatically be wrapped.