Allow Conjur longer to boot during ci
Since CI instances were upgraded to 20.04, Conjur has been taking
longer to boot in docker and causing timeouts when setting up
environments for cucubmer. The cause of this increase will need
to be investigated, but to enable builds to continue this commit
increases the time the build will wait for Conjur to boot.

Related: conjurinc/ops#795