Prevent security fork builds from pushing to dockerhub
Currently the Jenkinsfile pushes to dockerhub on all master branch builds.
The problem is that security forks also have master branches so they also
push to the edge tag to dockerhub. However the security fork master branches
are not kept up to date so they push old versions to edge.

This commit modifies the stage conditional to ensure that only builds
for the master branch in the main repo push to dockerhub.

Related: conjurinc/ops#790