Fix dev/start bug created by docker-compose 2
Previously, the code that verified Conjur server was up looked like this:

  output=$(docker-compose exec client conjur authn login -u admin -p $api_key |
    grep "Failed to open TCP connection" | wc -l)
  if [ $output -eq 0 ]; then

This assumes that when the docker-compose cmd fails, it prints its error
to stdout.  And this is true in the current docker-compose 1.x.  But if
you upgrade docker desktop it also upgrades you to docker-compose 2
(beta), which fixes this un-unix-y behavior by printing the error to
stderr.

This breaks the code above, because it thinks conjur is up when it's
not, and plows ahead.  But the error you see is seemingly unrelated:
It's just a prompt asking you to login to conjur as "alice".

Even worse, there is a race condition.  Sometimes conjur comes up fast,
and if it does, everything will work fine even though the "wait for
conjur" code is broken.  So there's some randomness just to add to the
fun.

The fix was easy once we figured out the root cause: Just redirect the
docker-compose commands stderr back to stdout.