Fix GKE CI flakiness
We believe the GKE flakiness was caused by the break statement which
this commit removes.  Essentially, we were assuming that there was only
one "expected" error that could happen, and that anything else was a
"real" error that we should mark as a failure.  However, if the pod was
not up yet (which is another exepected error), our code would stop
retrying.

This commit adds the "pod not up" error to the list of exepected ones,
and also keeps retrying the full 20 times no matter what, so that the
code won't be fragile to other expected errors we don't know about, or
to changes in the string messages of those errors.  This is safe because
we only retry a limited number of times anyway.