Fix another GKE flake
We believe there is a GKE timing problem causing flaky failures.  In
particular, prior to this commit, we were not waiting to ensure the
conjur_cli pod was up before copying policy templates to it.  So we are
adding:

  kubectl wait --for=condition=Ready "pod/$cli_pod" --timeout=5m

to the copyConjurPolicies function.  I've also gone through the entire
file to ensure that we alway wait after retrieving a pod name, before
taking any action against that pod.  I've added comments where this
waiting is not needed, if it's already been done in a previous step.