Push version tags to internal docker registry
Currently the images pushed to the internal registry are tagged
only by sha (despite the code comment to the contrary). This makes using
images from the internal registry difficult unless you happen to know
the sha you need.

This commit changes the internal registry tag format from ${SHA} to
${version}-${SHA}.

It also adds a push to the section that creates the version tags in
dockerhub so they are also created in the internal registry.

Related: conjurinc/ops#823