Initial PoC work to enable Host schema validation
This commit includes a change that allows an authenticator to tie into Conjur Host resource validation
lifecycle. This enables authenticators to define their own host annotation schema validation by including
a class called "Validations" with a method called "validate_host", which is run when a host is loaded from
policy.