Write to audit in the controller instead of in Authenticate class
The OIDC & GCP authenticators start their work with updating the authenticator
input with the username that is retrieved from the JWT token. If this operation
fails then we do not proceed in to the `Authenticate` class and thus the audit
message is not written. To overcome this, we move the audit logic to the controller