Integrate `public-keys` variable into full JWT Authentication flow
When Conjur is unable to reach remote JWKS URI endpoint public-keys variable can be used in order to provide static JWK Set to the authenticator.
Variable value is a JSON object contains two fields: type and value. The type filed should be set to "jwks" string and the value field should be set to a relevant JWKS set.