Clean up conjur server resources during tests
Previously conjur server processes spawn by rspec were
killed after the test, but the tests didn't wait for the process
to end. This meant that the shutdown process for prior
conjur servers was interacting with subsequent tests.

We added better mechanisms for running a test conjur
server in the PID tests. So we use it in the other tests
in this commit.