Fixes failing server_spec tests
It appears when these tests are run locally, the IO object is removed by an external process.  This change allows the tests to
pass locally. It catches ALL IO errors, which isn't ideal. I believe there's a deeper issue in either the way this is tested or the
way this is implemented.