Adds an internal service to provider Provider list to the UI
This commit includes an a rework of the authn-local socket server to enable
a secondary local socket (intended for Conjur UI) to deliver the list of available
OIDC Providers to the UI.

This work is a temporary stopgap. It will be removed when partial replication (Conjur
Enterprise) has been completed.