Add OIDC logout endpoint
Defines a new API endpoint, /authn-oidc/:account/:service_id/logout, that
revokes a user's refresh tokens, and returns to the user a URI to the OIDC
provider's session termintion endpoint.

When a user follows this URI, their browser session is cleared, and they are
sent to the configured post-logout redirect.