Allow Provider endpoint to handle incorrect OIDC provider endpoints
This commit addresses an issue discovered while testing MFA on Conjur UI. Previously, if any of the OIDC
provider's address is unreachable, the provider endpoint would cause an exception, which would cause the
login page to fail. This commit allows the provider endpoint to work even if the authenticator is incorrectly
configured.