Allows PKCE verification to be optional
This commit updates the PKCE feature to be on by default. The non-PKCE workflow
will be removed in a future release.

This change allows PKCE to be optional for OIDC code authentication. This is required
because the version of Keycloak we use for integration testing does not support PKCE.

Upgrading Keycloak will be handled in the near future.