Update cucumber RestHelpers to support parallel tests
The prior PR https://github.com/cyberark/conjur/pull/2818
updated our cucumber tests to use multiple processes,
running tests in parallel. This required parameterizing
several of the test resources to support two parallel
sets of tests.

This rest helper was missed in the prior PR and was
using only one environment's authn-local socket
rather than the socket for the given test environment.
This caused some tests to fail depending on the
environment in which they ran.