Fixes issue with conjurctl retrieve-key
Previously in order to determine if the corresponding
rake task was successful or not we checked if stderr
was empty. This is unreliable and deoesn't work
if certain rake logging levels are set because they
print warnings to stderr. Instead we now just check
the exit status of the rake task.