Update test dockerfile to be able to build locally
A recent commit c526c4b re-added the ruby-debug-ide
gem back into development dependencies. There was a required
update to run `bundle install` twice due to an issue with this
gem. This commit updates the Dockerfile.test so that it can
also be built locally/bypass this issue which only seems to
occur locally, not in CI.